Implement property path in YamlUtility
